--- app/controllers/admin_controller.rb
@@ -1,5 +1,0 @@
-class AdminController < ApplicationController
-  def index
-
-  end
-end
