--- app/controllers/invoice_items_controller.rb
@@ -1,2 +1,0 @@
-class InvoiceItemsController < ApplicationController
-end
