+++ app/controllers/merchants/dashboard_controller.rb
@@ -1,3 +1,7 @@
-class Merchants::DashboardController < ApplicationController
-  
+module Merchants
+  class DashboardController < ApplicationController
+    def index
+      @merchant = Merchant.find(params[:merchant_id])
+    end
+  end
 end
