+++ app/controllers/merchants/invoice_items_controller.rb
@@ -1,0 +1,4 @@
+module Merchants
+  class InvoiceItemsController < ApplicationController
+  end
+end
