--- app/controllers/merchants_controller.rb
@@ -1,7 +1,0 @@
-class MerchantsController < ApplicationController
-
-  def dashboard
-    @merchant = Merchant.find(params[:id])
-  end
-
-end
